requirements.txt
Django==4.1.5
aiohttp==3.8.3
python-dotenv==0.21.1
